Update entities - change id type to integer

+++ src/main/java/springtraining/luuquangbookmanagement/repositories/UserRepository.java
@@ -3,8 +3,8 @@ package springtraining.luuquangbookmanagement.repositories;
 import org.springframework.data.jpa.repository.JpaRepository;
 import springtraining.luuquangbookmanagement.repositories.entities.User;
 
-public interface UserRepository extends JpaRepository<User, Long> {
+public interface UserRepository extends JpaRepository<User, Integer> {
     User findByEmail(String email);
 
-    User findById(long id);
+    User findById(int id);
 }
